+++ modules/tarz/src/concatC.ts
@@ -4,7 +4,7 @@
 
 import {Action, isList, isNil, List, Nil} from '@action-land/core'
 import {CurriedFunction2, curry2} from 'ts-curry'
-import {CommandFunction} from './CommandFunction'
+import {CommandFunction} from './commandFunction'
 
 export const concatC = <State>(
   ...t: Array<CommandFunction<State>>
@@ -24,7 +24,7 @@ export const concatC = <State>(
       return result.length === 0
         ? Nil()
         : result.length === 1
-          ? result[0]
-          : List(...result)
+        ? result[0]
+        : List(...result)
     }
   )