+++ modules/tarz/src/concatR.ts
@@ -2,7 +2,7 @@
  * Created by tushar on 25/06/18
  */
 import {CurriedFunction2, curry2} from 'ts-curry'
-import {ReducerFunction} from './ReducerFunction'
+import {ReducerFunction} from './reducerFunction'
 
 /**
  * Concatenates 2 or more ReducerFunction(s) and returns a new ReducerFunction.