+++ modules/core/CHANGELOG.md
@@ -3,6 +3,17 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="4.0.1"></a>
+## [4.0.1](https://github.com/action-land/action-land/compare/v4.0.0...v4.0.1) (2018-08-22)
+
+
+### Bug Fixes
+
+* **action-type:** it should accept actions which have type as number ([a9f3ad5](https://github.com/action-land/action-land/commit/a9f3ad5))
+
+
+
+
 <a name="2.0.1"></a>
 ## [2.0.1](https://github.com/action-land/action-land/compare/v2.0.0...v2.0.1) (2018-07-26)